langchain/examples/python/requirements.txt
langchain==1.0.1
langchain-openai==1.0.1
langchain-azure-ai==0.1.8
azure-identity==1.25.1
python-dotenv==1.1.1